urlshortener/requirements.txt
asgiref==3.4.1
attrs==21.2.0
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0
